requirements.txt
streamlit==1.37.0
pandas==2.2.2
plotly==5.22.0
pymysql==1.1.1
streamlit-tags==1.2.5
Pillow==10.4.0
pdfplumber==0.11.1  # <-- New library
python-dotenv